moved dmrl tests into dmrl model folder and added checker for reqs being installed. Also fixed bug in PWLearningSampler
