+++ cornac/models/dmrl/test_distance_calc.py
