+++ cornac/models/dmrl/test_transformervision.py
@@ -3,13 +3,28 @@ Tests for the TransformersVisionModality class. In order to run this test please
 insert url_to_beach1, url_to_beach2, url_to_cat in the get_photos method. Use
 your favorite beach and cat photos and check the similarity scores.
 """
-import torch
-import unittest
-from PIL import Image
-from sentence_transformers import util
 
-from cornac.data.transformer_vision import TransformersVisionModality
-import requests
+# add a checker to make sure all requirements needed in the imports here are really present.
+# if they are missing skip the respective test
+# If a user wants to un these please run: pip install -r cornac/models/dmrl/requirements.txt
+
+try:
+    import torch
+    import unittest
+    from PIL import Image
+    from sentence_transformers import util
+
+    from cornac.models.dmrl.transformer_vision import TransformersVisionModality
+    import requests
+    run_dmrl_test_funcs = True
+
+except ImportError:
+    run_dmrl_test_funcs = False
+
+def skip_test_in_case_of_missing_reqs(test_func):
+  test_func.__test__ = run_dmrl_test_funcs  # Mark the test function as (non-)discoverable by unittest
+  return test_func
+
 
 # Please insert valid urls here to two beach photos and one cat photo
 beach_urls = ["url_to_beach1",
@@ -38,6 +53,7 @@ class TestTransformersVisionModality(unittest.TestCase):
         self.ids = [0, 1]
         self.modality = TransformersVisionModality(images=self.images, ids=self.ids, preencode=True)
 
+    @skip_test_in_case_of_missing_reqs
     @unittest.skipIf("url_to_beach1" in beach_urls, "Please insert a valid url to download 2 beach and one cat photo")
     def test_transform_image_to_tensor(self):
         """
@@ -48,6 +64,7 @@ class TestTransformersVisionModality(unittest.TestCase):
         assert image_tensor_batch.shape[0:2] == torch.Size((3, 3)) # 3 images with 3 channels each
         assert image_tensor_batch.shape[2:] == torch.Size(self.modality.image_size)
 
+    @skip_test_in_case_of_missing_reqs
     @unittest.skipIf("url_to_beach1" in beach_urls, "Please insert a valid url to download 2 beach and one cat photo")
     def test_encode_all_images(self):
         """
@@ -57,7 +74,8 @@ class TestTransformersVisionModality(unittest.TestCase):
         assert isinstance(self.modality.features, torch.Tensor)
         assert self.modality.features.shape[0] == len(self.images)
         assert self.modality.features.shape[1] == 1000
-    
+
+    @skip_test_in_case_of_missing_reqs
     @unittest.skipIf("url_to_beach1" in beach_urls, "Please insert a valid url to download 2 beach and one cat photo")
     def test_encoding_quality(self):
         """
